refactor: split ModalPartida into smaller components
