+++ components/DeleteAlertDialog.js
@@ -1,0 +1,48 @@
+import {
+    AlertDialog,
+    AlertDialogBody,
+    AlertDialogFooter,
+    AlertDialogHeader,
+    AlertDialogContent,
+    AlertDialogOverlay,
+    Button,
+} from "@chakra-ui/react";
+
+export default function DeleteAlertDialog({
+    isOpen,
+    cancelRef,
+    onCancel,
+    onConfirm,
+    modalRef,
+}) {
+    return (
+        <AlertDialog
+            isOpen={isOpen}
+            leastDestructiveRef={cancelRef}
+            onClose={onCancel}
+            isCentered
+            portalProps={{ containerRef: modalRef }}
+        >
+            <AlertDialogOverlay>
+                <AlertDialogContent>
+                    <AlertDialogHeader fontSize="lg" fontWeight="bold">
+                        Eliminar partida
+                    </AlertDialogHeader>
+
+                    <AlertDialogBody>
+                        ¿Estás seguro de que deseas eliminar la partida?
+                    </AlertDialogBody>
+
+                    <AlertDialogFooter>
+                        <Button ref={cancelRef} onClick={onCancel}>
+                            Cancelar
+                        </Button>
+                        <Button colorScheme="red" onClick={onConfirm} ml={3}>
+                            Aceptar
+                        </Button>
+                    </AlertDialogFooter>
+                </AlertDialogContent>
+            </AlertDialogOverlay>
+        </AlertDialog>
+    );
+}
